# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/patients": {
      "post": {
        "operationId": "PatientsController_createPatient",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePatientDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Patient profile created successfully"
          }
        },
        "summary": "Create a patient profile",
        "tags": [
          "Patients"
        ]
      },
      "get": {
        "operationId": "PatientsController_getAllPatients",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of all patients"
          }
        },
        "summary": "Get all patients",
        "tags": [
          "Patients"
        ]
      }
    },
    "/patients/{userId}": {
      "get": {
        "operationId": "PatientsController_getPatientProfile",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "Patient user ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Patient profile object"
          }
        },
        "summary": "Get patient profile by user ID",
        "tags": [
          "Patients"
        ]
      },
      "put": {
        "operationId": "PatientsController_updatePatientProfile",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "Patient user ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePatientProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Patient profile updated successfully"
          }
        },
        "summary": "Update patient profile",
        "tags": [
          "Patients"
        ]
      }
    },
    "/medicines": {
      "get": {
        "operationId": "MedicinesController_listMedicines",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of medicines"
          }
        },
        "summary": "List all medicines",
        "tags": [
          "Medicines"
        ]
      }
    },
    "/orders": {
      "post": {
        "operationId": "OrdersController_createOrder",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order added to cart successfully"
          }
        },
        "summary": "Add a medicine to cart",
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/cart/{userId}": {
      "get": {
        "operationId": "OrdersController_getCartOrdersByUserId",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get current cart for a user",
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/place/{userId}": {
      "post": {
        "operationId": "OrdersController_placeCartOrdersByUserId",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Place order for all items in the cart",
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/history/{userId}": {
      "get": {
        "operationId": "OrdersController_getPlacedOrdersByUserId",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get order history for a user",
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/cart/{orderId}": {
      "put": {
        "operationId": "OrdersController_updateCartOrder",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCartOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update quantity of an order in cart",
        "tags": [
          "Orders"
        ]
      },
      "delete": {
        "operationId": "OrdersController_removeCartOrder",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Remove an order from cart",
        "tags": [
          "Orders"
        ]
      }
    },
...
    "/medical-records/doctor/{doctorId}": {
      "get": {
        "operationId": "MedicalRecordsController_getRecordsByDoctorId",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "doctorId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of doctor medical records"
          }
        },
        "summary": "Get medical records by doctor ID",
        "tags": [
          "Medical Records"
        ]
      }
    },
    "/medical-records/{patientId}": {
      "get": {
        "operationId": "MedicalRecordsController_getRecordsByPatientId",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of patient medical records"
          }
        },
        "summary": "Get medical records by patient ID",
        "tags": [
          "Medical Records"
        ]
      }
    },
    "/medical-records": {
      "post": {
        "operationId": "MedicalRecordsController_createRecord",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "User role (admin, doctor, patient, frontdesk)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMedicalRecordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Medical record created successfully"
          }
        },
        "summary": "Create a new medical record",
        "tags": [
          "Medical Records"
        ]
      }
    },
...
  },
  "info": {
    "title": "MEDBITS Healthcare API",
    "description": "REST API for the MEDBITS Healthcare Management System.\n\n**Authentication**: Pass the user role in the `role` request header.\n\nValid roles: `patient` | `doctor` | `frontdesk` | `admin`",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "role": {
        "type": "apiKey",
        "in": "header",
        "name": "role",
        "description": "User role for RBAC"
      }
    },
    "schemas": {
...
      "CreatePatientDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "dob": {
            "type": "string",
            "example": "1990-01-01"
          },
          "gender": {
            "type": "string",
            "example": "Male"
          },
          "bloodGroup": {
            "type": "string",
            "example": "O+"
          },
          "phone": {
            "type": "string",
            "example": "9876543210"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          },
          "guardianName": {
            "type": "string",
            "example": "Jane Doe"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "dob",
          "gender",
          "bloodGroup",
          "phone",
          "email"
        ]
      },
      "UpdatePatientProfileDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "dob": {
            "type": "string",
            "example": "1990-01-01"
          },
          "gender": {
            "type": "string",
            "example": "Male"
          },
          "bloodGroup": {
            "type": "string",
            "example": "O+"
          },
          "phone": {
            "type": "string",
            "example": "9876543210"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com"
          },
          "guardianName": {
            "type": "string",
            "example": "Jane Doe"
          }
        }
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "PAT001"
          },
          "medicineId": {
            "type": "string",
            "example": "MED001"
          },
          "quantity": {
            "type": "number",
            "example": 2
          }
        },
        "required": [
          "userId",
          "medicineId",
          "quantity"
        ]
      },
      "UpdateCartOrderDto": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "quantity"
        ]
      },
...
      "CreateMedicalRecordDto": {
        "type": "object",
        "properties": {
          "doctorId": {
            "type": "string",
            "example": "DOC001"
          },
          "patientId": {
            "type": "string",
            "example": "PAT001"
          },
          "type": {
            "type": "string",
            "example": "consultation",
            "enum": [
              "consultation",
              "treatment",
              "lab"
            ]
          },
          "doctorName": {
            "type": "string",
            "example": "Dr. Smith"
          },
          "specialization": {
            "type": "string",
            "example": "Cardiology"
          },
          "date": {
            "type": "string",
            "example": "2026-05-10"
          },
          "consultationNote": {
            "type": "string",
            "example": "Patient complains of chest pain."
          },
          "medicines": {
            "type": "string",
            "example": "Aspirin 75mg"
          },
          "followUp": {
            "type": "string",
            "example": "Review after 1 week"
          },
          "followUpDate": {
            "type": "string",
            "example": "2026-05-17"
          },
          "appointmentId": {
            "type": "string",
            "example": "APT001"
          },
          "tests": {
            "type": "string",
            "example": "ECG"
          },
          "lifestyle": {
            "type": "string",
            "example": "Avoid spicy food"
          },
          "diet": {
            "type": "string",
            "example": "Low sodium"
          },
          "duration": {
            "type": "string",
            "example": "15 mins"
          }
        },
        "required": [
          "doctorId",
          "patientId",
          "doctorName",
          "specialization"
        ]
      },
...
    }
  }
}